Fix for issue during plugin update (#7447)
