+++ changelog/update-6705-temp-fix-arg-issue
@@ -1,0 +1,5 @@
+Significance: patch
+Type: dev
+Comment: Temp fix by reverting controller constructor and moving code to new file
+
+
